+++ src/apis/domains/dashboard.ts
@@ -1,30 +1,10 @@
 import instance from "@/src/apis/instance";
-
-export interface CreateDashboardRequest {
-  title: string;
-  color: string;
-}
-
-export interface UpdateDashboardRequest {
-  title?: string;
-  color?: string;
-}
-
-export interface Dashboard {
-  id: number;
-  title: string;
-  color: string;
-  createdAt: string;
-  updatedAt: string;
-  createdByMe: boolean;
-  userId: number;
-}
-
-export interface GetDashboardList {
-  cursorId: number;
-  totalCount: number;
-  dashboards: Dashboard[];
-}
+import {
+  CreateDashboardRequest,
+  Dashboard,
+  GetDashboardList,
+  UpdateDashboardRequest,
+} from "@/src/apis/types";
 
 export const dashboardsApi = {
   create: async (body: CreateDashboardRequest) => {
