+++ src/apis/types/index.ts
@@ -1,0 +1,1 @@
+export * from "@/src/apis/types/dashboards";
